Introduce GL-based font/glyph cache, move text rendering to entirely GL code.
Much much faster, especially on color-intensive screens.

Details follow...

Existing rendering model:
Find runs of similar text, and for each do the following:
* Use TTF to create a new surface and render the text using the given fg/bg
* Create a openGL texture for this surface, send the surface data to the GPU
* Blit the texture to the screen
* Delete the texture, delete the surface

This is especially terrible one "diverse" screens of text that alternate
characters frequently, although isn't very efficient in other cases either.

This patch changes this to the following:

At startup/when font changes:
* Render all printable characters for all fonts into one combined texture
 (this texture then lives on GPU indefinitely, unless font changes)

When rendering text, for each character:
* Compute coordinates in texture for this character, append to texture array
* Compute coordinates on screen for this character, append to vertex array
* Store desired color information at end of color array

After entire screen is renderered, send these arrays to the GPU as a single op.

One could in theory make this faster with a custom text-rendering shader using
GLES 2.0 but given the rest of the code's use of 1.1 it seemed fit.
(And because this is most of the way there and is quite fast already)